CONSOLE-4705: Convert some files to TypeScript
convert these files from `@console/internal`
- build-logs.jsx
- configmap.jsx
- feedback-local.jsx
- impersonate-notifier.jsx
- kube-admin-notifier.jsx
- cloud-provider.js
- draggable-item-types.js
- k8s-watcher.js
- job.js
- status.js
