+++ frontend/public/components/feedback-local.ts
@@ -1,6 +1,7 @@
 import { useTranslation } from 'react-i18next';
+import { getReportBugLink } from '@console/internal/module/k8s/cluster-settings';
 
-export function useFeedbackLocal(reportBug) {
+export function useFeedbackLocal(reportBug: ReturnType<typeof getReportBugLink>) {
   const { t } = useTranslation();
 
   return {
