+++ frontend/public/components/utils/cloud-provider.ts
@@ -1,8 +1,10 @@
-export const cloudProviderID = (node) => {
+import { NodeKind } from '@console/internal/module/k8s';
+
+export const cloudProviderID = (node: NodeKind): string => {
   return node.spec.providerID ? node.spec.providerID.split('://')[0] : '';
 };
 
-export const cloudProviderNames = (providerNames) => {
+export const cloudProviderNames = (providerNames: string[]): string => {
   if (providerNames.length) {
     const displayNames =
       providerNames.length === 1 ? providerNames[0] : `Hybrid (${providerNames.join(' , ')})`;
