+++ frontend/public/components/utils/draggable-item-types.ts
