+++ frontend/public/module/k8s/job.ts
@@ -1,6 +1,7 @@
 import i18next from 'i18next';
+import { JobKind } from './types';
 
-export const getJobTypeAndCompletions = ({ spec }) => {
+export const getJobTypeAndCompletions = ({ spec }: JobKind) => {
   // if neither completions nor parallelism are defined, then it is a non-parallel job.
   if (!spec.completions && !spec.parallelism) {
     return { type: i18next.t('public~Non-parallel'), completions: 1 };
